refactor(agent-core): split core routing surface
